refactor: change the saving mechanism to an easier way

+++ recommence/Config.py
@@ -6,7 +6,7 @@ class CheckpointConfig:
     save_path: str
     staging_path: str | None = None
     data_file: str = 'data.pkl'
-    save_every_interval: int | None = None
+    save_every_interval: float | None = None
 
     no_fail: bool = False
